Lift the verification review-queue endpoints out of web_server
api/verification.py takes the /api/verification family: stream, entry,
config, play, compare-stream, approve, delete, clean-orphans - plus the
two history helpers and the shared play-session/duration helpers. the
quarantine review routes stay behind for now and import those two
helpers back.

function bodies are byte-identical; only the decorator changed and the
two rebindable boot globals (download_orchestrator, matching_engine)
became getters. routemap before/after is identical on method+path (1179
rows), test_client smoke passes on the moved routes, and the
verification test files (admin gate, orphan cleanup) pass untouched.
web_server: 37,271 -> 36,999 lines.
